Move client version info to clientversion from utils and version files
