+++ doc/release-process.txt
@@ -2,7 +2,7 @@
 
 * update (commit) version in sources
   bitcoin-qt.pro
-  src/version.h
+  src/clientversion.h
   share/setup.nsi
   doc/README*
