+++ src/clientversion.h
@@ -1,11 +1,15 @@
-#ifndef CLIENTVERSION_H
-#define CLIENTVERSION_H
+// Copyright (c) 2009-2014 The Bitcoin developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_CLIENTVERSION_H
+#define BITCOIN_CLIENTVERSION_H
 
 //
 // client versioning
 //
 
-// These need to be macros, as version.cpp's and bitcoin-qt.rc's voodoo requires it
+// These need to be macros, as clientversion.cpp's and neutron*-res.rc's voodoo requires it
 #define CLIENT_VERSION_MAJOR       1
 #define CLIENT_VERSION_MINOR       1
 #define CLIENT_VERSION_REVISION    2
@@ -16,4 +20,31 @@
 #define STRINGIZE(X) DO_STRINGIZE(X)
 #define DO_STRINGIZE(X) #X
 
-#endif // CLIENTVERSION_H
+/**
+ * neutrond-res.rc includes this file, but it cannot cope with real c++ code.
+ * WINDRES_PREPROC is defined to indicate that its pre-processor is running.
+ * Anything other than a define should be guarded below.
+ */
+
+#if !defined(WINDRES_PREPROC)
+
+#include <string>
+#include <vector>
+
+static const int CLIENT_VERSION =
+    1000000 * CLIENT_VERSION_MAJOR  ///
+    + 10000 * CLIENT_VERSION_MINOR  ///
+    + 100 * CLIENT_VERSION_REVISION ///
+    + 1 * CLIENT_VERSION_BUILD;
+
+extern const std::string CLIENT_NAME;
+extern const std::string CLIENT_BUILD;
+extern const std::string CLIENT_DATE;
+
+
+std::string FormatFullVersion();
+std::string FormatSubVersion(const std::string& name, int nClientVersion, const std::vector<std::string>& comments);
+
+#endif // WINDRES_PREPROC
+
+#endif // BITCOIN_CLIENTVERSION_H
