+++ src/crypter.cpp
@@ -147,7 +147,7 @@ bool EncryptAES256(const SecureString& sKey, const SecureString& sPlaintext, con
 
     // Verify key sizes
     if(sKey.size() != 32 || sIV.size() != AES_BLOCK_SIZE) {
-        printf("crypter EncryptAES256 - Invalid key or block size: Key: %d sIV:%d\n", sKey.size(), sIV.size());
+        LogPrintf("crypter EncryptAES256 - Invalid key or block size: Key: %d sIV:%d\n", sKey.size(), sIV.size());
         return false;
     }
 
@@ -179,7 +179,7 @@ bool DecryptAES256(const SecureString& sKey, const std::string& sCiphertext, con
 
     // Verify key sizes
     if(sKey.size() != 32 || sIV.size() != AES_BLOCK_SIZE) {
-        printf("crypter DecryptAES256 - Invalid key or block size\n");
+        LogPrintf("crypter DecryptAES256 - Invalid key or block size\n");
         return false;
     }
