+++ src/main.h
@@ -5,6 +5,7 @@
 #ifndef BITCOIN_MAIN_H
 #define BITCOIN_MAIN_H
 
+#include "clientversion.h"
 #include "bignum.h"
 #include "sync.h"
 #include "net.h"
