+++ src/makefile.osx
@@ -63,7 +63,7 @@ CFLAGS += -std=c++11 -Wall -Wextra -Wformat -Wformat-security -Wno-unused-parame
 
 OBJS= \
     obj/alert.o \
-    obj/version.o \
+    obj/clientversion.o \
     obj/checkpoints.o \
     obj/netbase.o \
     obj/addrman.o \
@@ -143,7 +143,7 @@ obj/txdb-leveldb.o: leveldb/libleveldb.a
 
 obj/build.h: FORCE
 	/bin/sh ../share/genbuild.sh obj/build.h
-version.cpp: obj/build.h
+clientversion.cpp: obj/build.h
 DEFS += -DHAVE_BUILD_INFO
 
 obj/%.o: %.cpp
