+++ src/qt/res/bitcoin-qt.rc
@@ -24,7 +24,7 @@ BEGIN
             VALUE "InternalName",       "Neutron-qt"
             VALUE "LegalCopyright",     "2009-2014 The Bitcoin developers, 2012-2014 The NovaCoin & PPCoin developers, 2014 The Neutron developers"
             VALUE "LegalTrademarks1",   "Distributed under the MIT/X11 software license, see the accompanying file COPYING or http://www.opensource.org/licenses/mit-license.php."
-            VALUE "OriginalFilename",   "neutron-qt.exe"
+            VALUE "OriginalFilename",   "Neutron-qt.exe"
             VALUE "ProductName",        "Neutron-Qt"
             VALUE "ProductVersion",     VER_PRODUCTVERSION_STR
         END
